requirements.txt
absl-py==2.2.1
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.37.23
botocore==1.37.23
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.1.8
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
Django==5.1.7
djangorestframework==3.16.0
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.1
future==1.0.0
greenlet==3.1.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.18
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
redis==5.2.1
s3transfer==0.11.4
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.1
sympy==1.13.1
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
triton==3.2.0
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
zope.interface==7.2
